Fix module name in aqua apidoc index (#1185)
* Fix module name in aqua apidoc index

As part of the combined documentation build sphinx emits a warning:
"WARNING: duplicate object description of qiskit, other instance in
apidoc/aqua, use :noindex: for one of them". This is caused by the
module name for the ibmq provider index incorrectly saying this
document is for the module qiskit; instead it should be
qiskit.aqua which is actually the module for the index. This commit
corrects the oversignt.

* Remove the module line